Move /whoami endpoint out of authentication routines
Previously location was merely a convenient place to put it at the time;
it is not directly related to authn itself.

New file is called endpoints/users.js because this is a user information
endpoint and anticipates more endpoints for /users/… in time.
